+++ src/main/java/com/sparta/project/controller/OrderController.java
@@ -5,8 +5,8 @@ import com.sparta.project.dto.common.ApiResponse;
 import com.sparta.project.dto.common.PageResponse;
 import com.sparta.project.dto.order.OrderCreateRequest;
 import com.sparta.project.dto.order.OrderResponse;
+import com.sparta.project.permission.PermissionValidator;
 import com.sparta.project.service.OrderService;
-import com.sparta.project.util.PermissionValidator;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
