+++ src/main/java/com/sparta/project/controller/StoreCategoryController.java
@@ -5,8 +5,8 @@ import com.sparta.project.dto.common.PageResponse;
 import com.sparta.project.dto.storecategory.StoreCategoryCreateRequest;
 import com.sparta.project.dto.storecategory.StoreCategoryResponse;
 import com.sparta.project.dto.storecategory.StoreCategoryUpdateRequest;
+import com.sparta.project.permission.PermissionValidator;
 import com.sparta.project.service.StoreCategoryService;
-import com.sparta.project.util.PermissionValidator;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
 import lombok.RequiredArgsConstructor;
@@ -16,15 +16,7 @@ import org.springframework.data.domain.Sort.Direction;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.data.web.SortDefault;
 import org.springframework.security.core.Authentication;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequiredArgsConstructor
