+++ src/main/java/com/sparta/project/controller/StoreController.java
@@ -6,8 +6,8 @@ import com.sparta.project.dto.common.ApiResponse;
 import com.sparta.project.dto.common.PageResponse;
 import com.sparta.project.dto.store.StoreResponse;
 import com.sparta.project.dto.store.StoreUpdateRequest;
+import com.sparta.project.permission.PermissionValidator;
 import com.sparta.project.service.StoreService;
-import com.sparta.project.util.PermissionValidator;
 import jakarta.validation.constraints.NotNull;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
