+++ src/main/java/com/sparta/project/controller/UserController.java
@@ -2,18 +2,12 @@ package com.sparta.project.controller;
 
 
 import com.sparta.project.domain.enums.Role;
-import com.sparta.project.dto.common.PageResponse;
-import com.sparta.project.dto.user.UserAdminResponse;
-import com.sparta.project.dto.user.UserBasicResponse;
-import com.sparta.project.dto.user.UserLoginRequest;
-import com.sparta.project.dto.user.UserResponse;
-import com.sparta.project.dto.user.UserSignupRequest;
 import com.sparta.project.dto.common.ApiResponse;
-import com.sparta.project.dto.user.UserUpdateRequest;
+import com.sparta.project.dto.common.PageResponse;
+import com.sparta.project.dto.user.*;
+import com.sparta.project.permission.PermissionValidator;
 import com.sparta.project.service.UserService;
-import com.sparta.project.util.PermissionValidator;
 import jakarta.validation.Valid;
-import java.util.Map;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -21,15 +15,9 @@ import org.springframework.data.domain.Sort.Direction;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.data.web.SortDefault;
 import org.springframework.security.core.Authentication;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.Map;
 
 @RestController
 @RequiredArgsConstructor
@@ -86,7 +74,7 @@ public class UserController {
     // 회원 정보 조회 (MANAGER, MASTER)
     @GetMapping("/{user_id}")
     public ApiResponse<UserResponse> getUserById(Authentication authentication,
-                                                      @PathVariable Long user_id) {
+                                                 @PathVariable Long user_id) {
         permissionValidator.checkPermission(authentication, Role.MANAGER.name(), Role.MASTER.name());
         UserResponse result = userService.getUserById(user_id, true);
         return ApiResponse.success(result);
