+++ src/main/java/com/sparta/project/permission/OrderValidator.java
@@ -1,27 +1,13 @@
-package com.sparta.project.util;
+package com.sparta.project.permission;
 
 import com.sparta.project.domain.Order;
 import com.sparta.project.domain.Store;
 import com.sparta.project.exception.CodeBloomException;
 import com.sparta.project.exception.ErrorCode;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.stereotype.Service;
+import org.springframework.stereotype.Component;
 
-import java.util.Arrays;
-
-@Service
-public class PermissionValidator {
-
-    public void checkPermission(Authentication authentication, String... roles) {
-        boolean hasPermission = Arrays.stream(roles)
-                .anyMatch(role -> authentication.getAuthorities()
-                                .contains(new SimpleGrantedAuthority(role)));
-        if (!hasPermission) {
-            throw new CodeBloomException(ErrorCode.FORBIDDEN_ACCESS);
-        }
-    }
-
+@Component
+public class OrderValidator {
 
     public void checkOrderApprovePermission(Order order, Long userId) {
         if (!isStoreOwner(order.getStore(), userId)) {
@@ -54,5 +40,4 @@ public class PermissionValidator {
             throw new CodeBloomException(ErrorCode.FORBIDDEN_ACCESS);
         }
     }
-
 }
