+++ src/main/java/com/sparta/project/permission/PermissionValidator.java
@@ -1,0 +1,22 @@
+package com.sparta.project.permission;
+
+import com.sparta.project.exception.CodeBloomException;
+import com.sparta.project.exception.ErrorCode;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.stereotype.Service;
+
+import java.util.Arrays;
+
+@Service
+public class PermissionValidator {
+
+    public void checkPermission(Authentication authentication, String... roles) {
+        boolean hasPermission = Arrays.stream(roles)
+                .anyMatch(role -> authentication.getAuthorities()
+                        .contains(new SimpleGrantedAuthority(role)));
+        if (!hasPermission) {
+            throw new CodeBloomException(ErrorCode.FORBIDDEN_ACCESS);
+        }
+    }
+}
