+++ src/main/java/com/sparta/project/service/OrderService.java
@@ -7,9 +7,9 @@ import com.sparta.project.dto.order.OrderMenuInfo;
 import com.sparta.project.dto.order.OrderResponse;
 import com.sparta.project.exception.CodeBloomException;
 import com.sparta.project.exception.ErrorCode;
+import com.sparta.project.permission.OrderValidator;
 import com.sparta.project.repository.OrderMenuRepository;
 import com.sparta.project.repository.OrderRepository;
-import com.sparta.project.util.PermissionValidator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -28,7 +28,7 @@ public class OrderService {
     private final StoreService storeService;
     private final OrderRepository orderRepository;
     private final OrderMenuRepository orderMenuRepository;
-    private final PermissionValidator permissionValidator;
+    private final OrderValidator orderValidator;
 
     @Transactional
     public String createOrder(Long userId, OrderCreateRequest request) {
@@ -67,7 +67,7 @@ public class OrderService {
     @Transactional
     public String approveOrder(String orderId, Long userId) {
         Order order = getOrderOrException(orderId);
-        permissionValidator.checkOrderApprovePermission(order, userId);
+        orderValidator.checkOrderApprovePermission(order, userId);
         order.approve();
         return order.getOrderId();
     }
@@ -76,14 +76,14 @@ public class OrderService {
     @Transactional
     public String cancelOrder(String orderId, Long userId) {
         Order order = getOrderOrException(orderId);
-        permissionValidator.checkOrderCancelPermission(order, userId);
+        orderValidator.checkOrderCancelPermission(order, userId);
         order.cancel(userId);
         return order.getOrderId();
     }
 
     public OrderResponse getOrderById(String orderId, long userId) {
         Order order = getOrderOrException(orderId);
-        permissionValidator.checkOrderOwner(userId, order.getUser().getUserId());
+        orderValidator.checkOrderOwner(userId, order.getUser().getUserId());
         return OrderResponse.from(order);
     }
 
@@ -111,7 +111,7 @@ public class OrderService {
         Store store = storeService.getStoreOrException(storeId);
 
         // store가 owner의 것이 맞는지 체크
-        permissionValidator.checkStoreOwnerPermission(store, ownerId);
+        orderValidator.checkStoreOwnerPermission(store, ownerId);
 
         // store 와 customer를 모두 만족하는 order
         return orderRepository
